+++ src/ping.h
@@ -61,6 +61,6 @@ struct _ping_data {
 /* Ping funtions */
 void ping_do (Netinfo * netinfo);
 void ping_stop (Netinfo * netinfo);
-void ping_foreach (Netinfo * netinfo, gchar * line, gint len, gpointer user_data);
+void ping_foreach (Netinfo * netinfo, gchar * line, gssize len, gpointer user_data);
 void ping_foreach_with_tree (Netinfo * netinfo, gchar * line, gint len, gpointer user_data);
 void ping_copy_to_clipboard (Netinfo * netinfo, gpointer user_data);
